Intent on ListView
Added an intent so whenever an item in the Listview is clicked, it will open the Editor activity and also it will change the title. If an items was clicked from the ListView, it will open with the title - Edit product, and if it was click the new product FAB, it will set the title Add a new product.
Also set the focusable option on the button to false, so I will be able to click on the items in the ListView