AlertDialog when Order button is pressed
An alert dialog will be displayed when the Order button is pressed and it will ask the user how many pieces of that specific product does he want to order.
If it clicks on the cancel button it will close the alert dialog.
If he clicks on the Order button, it will open the e-mail app to order the products