BUG Empty EditText EditorActivity
Before if the order EditText was empty and the user would had pressed on the minus or plus button, the app would had crashed. Now if the order EditText is empty, if the user pressed on the minus button, it will automatically set the value to 0 and if he presses on the plus button it will start incrementing.